requirements.txt
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
flatbuffers==1.12
gast==0.4.0
gevent==21.12.0
gunicorn
grpcio==1.46.3
h5py==3.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
# keras==2.9.0
# Keras-Preprocessing==1.1.2
libclang==14.0.1
llvmlite==0.38.1
Markdown==3.3.7
MarkupSafe==2.1.1
munkres==1.1.4
oauthlib==3.2.0
opt-einsum==3.3.0
pandas==1.4.2
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.25.1
python-speech-features==0.6
pytz==2022.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
# tensorboard==2.9.0
# tensorboard-data-server==0.6.1
# tensorboard-plugin-wit==1.8.1
# tensorflow-estimator==2.9.0
tensorflow-cpu==2.8.0
tensorflow-io-gcs-filesystem==0.26.0
Werkzeug==2.1.2
numpy
matplotlib
IPython
# torchaudio
librosa
soundfile
